Subiendo generación de factura

+++ database/migrations/2024_10_25_211307_add_imageurl_to_products_table.php
@@ -4,16 +4,18 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->string('ImageURL')->nullable()->after('Status'); // Añadir la columna ImageURL
-        });
+        // Verificar si la columna `ImageURL` ya existe
+        if (!Schema::hasColumn('products', 'ImageURL')) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->string('ImageURL')->nullable()->after('Status'); // Añadir la columna ImageURL solo si no existe
+            });
+        }
     }
 
     /**
@@ -21,8 +23,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('ImageURL'); // Eliminar la columna ImageURL
-        });
+        // Eliminar la columna `ImageURL` solo si existe
+        if (Schema::hasColumn('products', 'ImageURL')) {
+            Schema::table('products', function (Blueprint $table) {
+                $table->dropColumn('ImageURL'); // Eliminar la columna ImageURL
+            });
+        }
     }
 };
